backend/requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
docstring_parser==0.16
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.2
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
StrEnum==0.4.15
sympy==1.13.1
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.19.0
flask
python-dotenv
requests
sentence-transformers
supabase
openai
fastapi
uvicorn
python-dotenv
google-generativeai
supabase
